+++ MobileSDK/Sources/MobileSDK/Network/Requests/PayPalVault/PayPalVaultSetupTokenReq.swift
@@ -9,7 +9,6 @@
 struct PayPalVaultSetupTokenReq: Encodable {
     
     let gatewayId: String
-    let token: String
     let returnUrl = "sdk.ios.paypal://vault/success"
     let cancelUrl = "sdk.ios.paypal://vault/cancel"